Builds the Diagrammer function call.

Builds the Diagrammer function call.

Usage

buildCall(
  model = model,
  name = name,
  labels = labels,
  graph_options = list(overlap = "true", fontsize = "10"),
  node_options = list(shape = "box"),
  edge_options = list(color = "black"),
  ...
)

Arguments

model

A model fit object of class lavaan.

name

A string of the name of the plot.

labels

An optional named list of variable labels fit object of class lavaan.

graph_options

A named list of graph options for Diagrammer syntax. See for options here https://graphviz.org/docs/graph/

node_options

A named list of node options for Diagrammer syntax.

edge_options

A named list of edge options for Diagrammer syntax.

...

additional arguments to be passed to buildPaths

Value

A string specifying the path diagram for model