Extracts the paths from the lavaan model.

Extracts the paths from the lavaan model.

Usage

buildPaths(fit, coefs, sig, stand, covs, stars, digits, conf.int, edge_styles)

Arguments

fit

A model fit object of class lavaan.

coefs

whether or not to include significant path coefficient values in diagram

sig

significance level for determining what significant paths are

stand

Should the coefficients being used be standardized coefficients

covs

Should model covariances be included in the diagram

stars

a character vector indicating which parameters should include significance stars be included for regression paths, latent paths, or covariances. Include which of the 3 you want ("regress", "latent", "covs"), default is none.

digits

A number indicating the desired number of digits for the coefficient values in the plot

conf.int

whether or not to include confidence intervals of path coefficients

edge_styles

Whether or not to depict non-significant coefficients with dashed paths